Conflicts:
	Event/xAOD/xAODTrackingCnv/python/xAODTrackingCnvConfig.py
	InnerDetector/InDetConfig/python/InDetAmbiTrackSelectionToolConfig.py
	InnerDetector/InDetConfig/python/InDetConfigFlags.py
	InnerDetector/InDetConfig/python/TrackRecoConfig.py
	PhysicsAnalysis/DerivationFramework/DerivationFrameworkInDet/python/InDetToolsConfig.py
	Tracking/TrkConfig/python/TrkAmbiguityProcessorConfig.py
	Tracking/TrkConfig/python/TrkValToolsConfig.py

+++ Reconstruction/Jet/JetUncertainties/Root/UncertaintyHistogram.cxx
@@ -193,7 +193,7 @@ double UncertaintyHistogram::readHisto(const double var1, const double var2, con
     }
 
     // Check first dimension boundaries, always applicable
-    const float valX = checkBoundariesByBin(std::as_const(*m_histo).GetXaxis(),m_histo->GetNbinsX(),var1);
+    const double valX = checkBoundariesByBin(std::as_const(*m_histo).GetXaxis(),m_histo->GetNbinsX(),var1);
     if (m_nDim == 1)
     {
         switch (m_interpolate)
@@ -212,7 +212,7 @@ double UncertaintyHistogram::readHisto(const double var1, const double var2, con
     }
 
     // Check second dimension boundaries, if applicable
-    const float valY = checkBoundariesByBin(std::as_const(*m_histo).GetYaxis(),m_histo->GetNbinsY(),var2);
+    const double valY = checkBoundariesByBin(std::as_const(*m_histo).GetYaxis(),m_histo->GetNbinsY(),var2);
     if (m_nDim == 2)
     {
         // We need a 2D histogram for the projection calls
@@ -240,7 +240,7 @@ double UncertaintyHistogram::readHisto(const double var1, const double var2, con
     }
 
     // Check third dimension boundaries, if applicable
-    const float valZ = checkBoundariesByBin(std::as_const(*m_histo).GetZaxis(),m_histo->GetNbinsZ(),var3);
+    const double valZ = checkBoundariesByBin(std::as_const(*m_histo).GetZaxis(),m_histo->GetNbinsZ(),var3);
     
     switch (m_interpolate)
     {
